+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * Add new method :meth:`common.sort_variants`.
 * Add new method :meth:`pyvcf.VcfFrame.variants`.
 * Add new method :meth:`pymaf.MafFrame.variants`.
+* Add new method :meth:`pymaf.MafFrame.subset`.
 
 0.24.0 (2021-10-02)
 -------------------